+++ hyperspy/defaults_parser.py
@@ -138,10 +138,6 @@ class GUIs(t.HasTraits):
 
 
 class PlotConfig(t.HasTraits):
-    saturated_pixels = t.CFloat(0.,
-                                label='Saturated pixels (deprecated)',
-                                desc='Warning: this is deprecated and will be removed in HyperSpy v2.0'
-                                )
     # Don't use t.Enum to list all possible matplotlib colormap to
     # avoid importing matplotlib and building the list of colormap
     # when importing hyperpsy