+++ CHANGES.rst
@@ -128,7 +128,7 @@ migration of the IO plugins to `RosettaSciIO
   and ``metadata_keys`` keywords were removed. Use ``dataset_key``, ``dataset_path``
   and ``metadata_key`` instead.
 
-Data Visualisation
+Data Visualization
 ------------------
 
 - The :py:meth:`~._signals.signal2d.Signal2D.plot` keyword argument